requirements.txt
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
decorator==4.4.2
dialogflow==1.1.0
feedparser==6.0.2
Flask==1.1.2
fuzzywuzzy==0.18.0
gnewsclient==1.12
google-api-core==1.25.1
google-auth==1.24.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
levenshtein==0.12.0
MarkupSafe==1.1.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-telegram-bot==12.8
pytz==2020.5
requests==2.25.1
rsa==4.7
sgmllib3k==1.0.0
six==1.15.0
tornado==6.1
urllib3==1.26.3
Werkzeug==1.0.1
